requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.74.1
h11==0.13.0
idna==3.3
joblib==1.1.0
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
pydantic==1.9.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
threadpoolctl==3.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.1.1
uvicorn==0.17.5
